requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiohttp==3.9.1
requests==2.31.0
asyncio-throttle==1.0.2
beautifulsoup4==4.12.2
selenium==4.15.2
undetected-chromedriver==3.5.4
fake-useragent==1.4.0
trafilatura==1.6.4
newspaper3k==0.2.8
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1
Pillow==10.1.0
opencv-python==4.8.1.78
pytube==15.0.0
chromadb==0.4.18
openai>=1.6.1,<2.0.0
voyageai==0.2.1
ollama==0.1.7
replicate==0.24.0 
pandas==2.1.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4

fastapi-mail==1.4.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==41.0.7
pymilvus==2.4.4
marshmallow==3.19.0
environs==9.5.0
jinja2==3.1.2
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
setuptools<81

# PostgreSQL Vector Database Dependencies
asyncpg>=0.29.0                    # PostgreSQL async driver
psycopg2-binary>=2.9.9            # PostgreSQL sync driver (for migrations)

# Keep existing dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
openai>=1.3.0

# AI Services
anthropic>=0.7.0                   # If using Claude
requests>=2.31.0

# Web Scraping
beautifulsoup4>=4.12.0
playwright>=1.40.0
selenium>=4.15.0
lxml>=4.9.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
aiofiles>=23.2.0
python-magic>=0.4.27

# REMOVE Milvus dependencies (comment out or delete):
# pymilvus>=2.3.0                  # NOT NEEDED ANYMORE
# grpcio>=1.59.0                   # Milvus gRPC dependency
# protobuf>=4.25.0                 # Milvus protobuf dependency
